added join and leave endpoints
