+++ server/src/routes/rooms.routes.ts
@@ -7,15 +7,14 @@ import * as controller from "../controllers/rooms.controller";
 
 const router = Router();
 
-router.put("/", controller.createRoom)
+router.put("/", controller.createRoom);
 
-router.get("/:roomID", controller.joinRoom)
-router.put("/:roomID", controller.updateRoom)
-router.delete("/:roomID", controller.deleteRoom)
-router.get("/:roomID/users", controller.listUsers)
-router.put("/:roomID/users", controller.inviteUser)
-router.delete("/:roomID/users/:username", controller.removeUser)
-router.put("/:roomID/users/:username", controller.updateUser)
+router.put("/:roomID", controller.updateRoom);
+router.delete("/:roomID", controller.deleteRoom);
+router.get("/:roomID/users", controller.listUsers);
+router.put("/:roomID/users", controller.inviteUser);
+router.delete("/:roomID/users/:username", controller.removeUser);
+router.put("/:roomID/users/:username", controller.updateUser);
 
 // imported routers, users required to be part of the room to access
 router.use("/:roomID", confirmUserInRoom);
