+++ server/src/routes/user.routes.ts
@@ -14,7 +14,9 @@ router.put("/update", controller.update);
 router.get("/rooms", controller.listRooms);
 router.get("/rooms/:roomID", controller.getRoomDetails);
 router.delete("/rooms/:roomID", controller.removeRoomFromUser);
-router.put("/rooms/:roomID/invitation", controller.acceptRoomInvite)
-router.delete("/rooms/:roomID/invitation", controller.declineRoomInvite)
+router.put("/rooms/:roomID/invitation", controller.acceptRoomInvite);
+router.delete("/rooms/:roomID/invitation", controller.declineRoomInvite);
+router.put("/rooms/:roomID/presence", controller.joinRoom);
+router.delete("/rooms/presence", controller.leaveRoom);
 
 export default router;
